feat: Introduce UniversalSearchDappItem component for improved Dapp rendering in UniversalSearch
