+++ packages/kit/src/views/UniversalSearch/components/SearchResultItems/index.ts
@@ -1,0 +1,1 @@
+export * from './UniversalSearchDappItem';
